+++ src/testFixtures/kotlin/com/widehouse/cafe/cafe/dto/CafeRequestFixture.kt
@@ -1,8 +1,5 @@
 package com.widehouse.cafe.cafe.dto
 
-class CafeRequestFixture {
-    companion object {
-        fun create() =
-            CafeRequest("test", "test", "desc", 1L)
-    }
+object CafeRequestFixture {
+    fun create() = CafeRequest("test", "test", "desc", 1L)
 }
