+++ src/testFixtures/kotlin/com/widehouse/cafe/cafe/dto/CafeResponseFixture.kt
@@ -2,12 +2,10 @@ package com.widehouse.cafe.cafe.dto
 
 import com.widehouse.cafe.cafe.model.Category
 
-class CafeResponseFixture {
-    companion object {
-        fun create(url: String = "test", name: String = "test", description: String = "desc", categoryId: Long = 1L) =
-            CafeResponse(url, name, description, Category(categoryId, ""))
+object CafeResponseFixture {
+    fun create(url: String = "test", name: String = "test", description: String = "desc", categoryId: Long = 1L) =
+        CafeResponse(url, name, description, Category(categoryId, ""))
 
-        fun from(index: Int, categoryId: Long = 1L) =
-            create("test$index", "test$index", "desc$index", categoryId)
-    }
+    fun from(index: Int, categoryId: Long = 1L) =
+        create("test$index", "test$index", "desc$index", categoryId)
 }
